🍌 Split myst-template functionality out of jtex
