+++ packages/jtex/src/index.ts
@@ -1,10 +1,4 @@
-import JTex from './jtex';
-export * from './download';
 export * from './export';
-export * from './frontmatter';
-export * from './session';
-export * from './types';
-export * from './validators';
 export * from './imports';
-
-export default JTex;
+export * from './jtex';
+export * from './types';
