+++ packages/myst-templates/package.json
@@ -9,6 +9,7 @@
   "module": "dist/esm/index.js",
   "types": "dist/types/index.d.ts",
   "files": [
+    "src",
     "dist"
   ],
   "exports": {
@@ -43,14 +44,32 @@
     "url": "https://github.com/executablebooks/mystjs/issues"
   },
   "dependencies": {
-    "myst-frontmatter": "^0.0.3"
+    "chalk": "^4.1.2",
+    "commander": "^8.3.0",
+    "crypto": "^1.0.1",
+    "js-yaml": "^4.1.0",
+    "myst-cli-utils": "^0.0.7",
+    "myst-frontmatter": "^0.0.3",
+    "node-fetch": "^2.6.7",
+    "pretty-hrtime": "^1.0.3",
+    "simple-validators": "^0.0.2",
+    "unzipper": "^0.10.11"
   },
   "devDependencies": {
+    "@types/jest": "^28.1.6",
+    "@types/node-fetch": "^2.6.2",
+    "@types/nunjucks": "^3.2.1",
+    "@types/sanitize-html": "^2.6.2",
+    "@types/unzipper": "^0.10.5",
+    "esbuild": "^0.15.12",
+    "eslint": "^8.21.0",
+    "eslint-config-curvenote": "latest",
+    "jest": "^28.1.3",
+    "memfs": "^3.4.7",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.6.1",
+    "prettier": "latest",
     "rimraf": "^3.0.2",
     "ts-jest": "^28.0.7",
-    "ts-loader": "^9.2.8",
-    "typescript": "^4.6.3"
+    "typescript": "latest"
   }
 }
