+++ packages/myst-templates/src/download.ts
@@ -1,13 +1,12 @@
-import { createHash } from 'crypto';
 import fs, { createReadStream, createWriteStream, mkdirSync } from 'fs';
-import yaml from 'js-yaml';
-import type { TemplateYmlListResponse, TemplateYmlResponse } from 'myst-templates';
-import fetch from 'node-fetch';
-import unzipper from 'unzipper';
 import { dirname, join, parse } from 'path';
-import { validateUrl } from 'simple-validators';
-import type { ISession } from './types';
+import { createHash } from 'crypto';
+import yaml from 'js-yaml';
 import { createGitLogger, makeExecutable } from 'myst-cli-utils';
+import fetch from 'node-fetch';
+import { validateUrl } from 'simple-validators';
+import unzipper from 'unzipper';
+import type { TemplateYmlListResponse, TemplateYmlResponse, ISession } from './types';
 
 export const TEMPLATE_FILENAME = 'template.tex';
 export const TEMPLATE_YML = 'template.yml';
