+++ packages/myst-templates/src/session.ts
