+++ packages/myst-templates/src/validators.ts
@@ -8,14 +8,6 @@ import {
   validateLicenses,
   validatePageFrontmatter,
 } from 'myst-frontmatter';
-import type {
-  TemplateDocDefinition,
-  TemplateOptionDefinition,
-  TemplatePartDefinition,
-  TemplateStyles,
-  TemplateYml,
-} from 'myst-templates';
-import { TemplateOptionTypes } from 'myst-templates';
 import {
   defined,
   incrementOptions,
@@ -32,6 +24,14 @@ import {
   validationError,
 } from 'simple-validators';
 import type { ValidationOptions } from 'simple-validators';
+import type {
+  TemplateDocDefinition,
+  TemplateOptionDefinition,
+  TemplatePartDefinition,
+  TemplateStyles,
+  TemplateYml,
+} from './types';
+import { TemplateOptionTypes } from './types';
 
 /** Validate that input is an existing file name
  *
